pyproject.toml
[project]
name = "fintech-rag-system"
version = "0.1.0"
description = "Financial Document RAG System with LLM"
authors = [
    {name = "xuchenyang6688",email = "xuchenyang6688@163.com"}
]
license = {text = "MIT"}
requires-python = "^3.10"
dependencies = [
    "langchain[schema] (>=1.2.7,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "httpx-sse (>=0.4.3,<0.5.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
